requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
beautifulsoup4==4.13.3
bidict==0.23.1
bs4==0.0.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cmudict==1.0.32
colorama==0.4.6
cryptography==44.0.2
Cython==3.0.12
defusedxml==0.7.1
Distance==0.1.3
dj-rest-auth==7.0.1
Django==4.2.20
django-allauth==65.6.0
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
emoji==1.2.0
exceptiongroup==1.2.2
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.3.0
h11==0.14.0
hangul-jamo==1.0.1
huggingface-hub==0.29.3
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.1.0
jamo==0.4.1
Jinja2==3.1.6
joblib==1.4.2
kollocate==0.0.2
koparadigm==0.10.0
kss==6.0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==2.6.3
numpy==1.26.4
oauthlib==3.2.2
packaging==24.2
pecab==1.0.8
pillow==11.1.0
pluggy==1.5.0
pyarrow==19.0.1
pybind11==2.9.2
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
PyJWT==2.9.0
pytest==8.3.5
python-decouple==3.8
python-dotenv==1.1.0
python-mecab-ko==1.3.7
python-mecab-ko-dic==2.1.1.post2
python3-openid==3.2.0
PyYAML==6.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.10.1
sentence-transformers==4.0.2
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.5.6
soupsieve==2.6
sqlparse==0.5.3
starlette==0.46.1
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.1
tomli==2.2.1
torch==2.6.0
tossi==0.3.1
tqdm==4.67.1
transformers==4.50.3
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
Unidecode==1.3.8
urllib3==2.3.0
uv==0.6.12
uvicorn==0.34.0
Whoosh==2.7.4
xlrd==1.2.0
zipp==3.21.0
